Moved RobotLikeComponent to a different package. I'm not sure why frc.robot.core is desirable, but I'll mimic that for now.
